Moved hardware/thermometer/__init__.py to hardware/thermometer/core.py
