+++ backend/hardware/devicelist.py
@@ -1,7 +1,7 @@
 import hardware.stirring.core as stirrer
 import hardware.reagentdispenser.core as rd
 import hardware.temperaturecontroller.core as tc
-import hardware.thermometer as thermometer
+import hardware.thermometer.core as thermometer
 import hardware.gpiochip.core as gpiochip
 from config import microlabConfig as config
 import yaml
